Finished EnemyAI!!!
The enemy can now seek the players final plotted point on the x/y axis. After reaching this point, the enemy with scan for the player target again and if the player is within range, will seek the player. Added a circle collider2d to the enemy temporarily. The enemy was getting stuck on the corners of obstacles and this worked as a solution.