Created Signs With Text!
Created signs with customizable text. The signs are diamond shaped obstacles that the player can trigger with keycode.E. For this to work I created a dialog box gameobject along with a text game object, and added a signs script I had from a previous project. Can also use these to trigger events. Might be something cool to consider for the future.