Lighting/Highlighting For Interactable Objects
Created a background and added lighting for the scene. To do this, I downloaded the Universal RP package and changed the graphics under project settings to URP Lights (Universal Render Pipeline Assets). Following, I created the background and made it a darker shade(this will influence the interactable object highlighting) and made the order in layer -5 to ensure all game objects will show. Additionally, I changed the material for every obstacle, player and enemy to Sprite-Lit-Default from the default material (this causes all of the objects to go black, requiring a light source). Finally, I added a spotlight over the player so it can show objects it will run into, as well as a Freeform Light on the player and all interactable objects (this causes a highlight effect but will be adjusted is the background color is adjusted as well).