Object Highlighting Scripts
Finished interactable object highlighting. Created and added a script called HighlightTrigger that will enable the the Freeform light to turn on when a game object with the tag "Player" is within range, and it will turn back off when the player is out of range. To do this, I used the package:

using UnityEngine.Rendering.Universal;

This allows us to call upon the Light2D component/syntax, which controls the Freeformlight. Following I just declared Light2D as light and then told Visual studio that light = GetComponent<Light2D>();

AKA the light = FreeformLight

Moreover, I took 'light' and used it within my if statement to to trigger the light to turn on and off whenever the game object with "Player" tag is within range.

Finally, I made sure to add this script to the freeform light, and to turn off the 'Light 2D' C# scripts that are automatically applied to the game objects with FreeformLight to make sure they will only turn on when the player is in range.