Enemy Patrolling Adjustments
Added a new enemy patrolling script and adjusted the enemy patrolling points. Now the enemy is able to patrol between different points at a set speed and a set wait time. When the player is within range, the enemy will stop patrolling and begin to chase the player. Once the player is out of range, the enemy will return to its patrol point coordinate. Still needing to fix a bug with the patrolling where the enemy does not avoid obstacles, but instead will run into them and slide to the point it needs to get to.